Change text to float

+++ app/src/main/java/id/themanusia/projekakhirmikro/Dht22.kt
@@ -3,12 +3,12 @@ package id.themanusia.projekakhirmikro
 import com.google.firebase.database.PropertyName
 
 data class Dht22(
-    @PropertyName("temperature") val temperature: Double = 0.0,
-    @PropertyName("humidity") val humidity: Double = 0.0,
-    @PropertyName("heatindex") val heatIndex: Double = 0.0,
+    @PropertyName("temperature") val temperature: Float = 0.0f,
+    @PropertyName("humidity") val humidity: Float = 0.0f,
+    @PropertyName("heatindex") val heatIndex: Float = 0.0f,
     @PropertyName("random") val random: Int = 0,
     ) {
-    constructor() : this(0.0,0.0,0.0, 0)
+    constructor() : this(0.0f,0.0f,0.0f, 0)
 
     override fun toString(): String {
         return "peler gede: $temperature, $humidity, $heatIndex, $random"
